Add src/util/helpers and move more types into types.ts
